+++ NEWS.md
@@ -1,5 +1,8 @@
 # plu 0.1.1
 
+* Allow braces to be used within words.
+  * "cact{us|i}" -> "cactus", "cacti"
+  * "antenna{|e}" -> "antenna", "antennae"
 * Remove cross-reference to **glue** to avoid CRAN NOTE.
 * Add AGID copyright information to package documentation (`?plu`).
 * Minor speed improvements to regular pluralization rules.