+++ NEWS.md
@@ -1,3 +1,5 @@
+# plu (development version)
+
 # plu 0.1.1
 
 * Allow braces to be used within words.